HELLO FROM 42K FEET IN THE AIR!!
there is wifi in this plane SOMEHOW. all this time crammed in a tube finally got me to advance the scripting a little bit.

now that i know what to manipulate to change the texutureRects, i now have to find out how to make an array of the possible choices the opponent player has, then randomize them and trigger them when any of the user's 3 buttons are clicked, and behaviour of course.